requirements.txt
streamlit
pandas
requests
tensorflow
Pillow
numpy
python-dotenv